requirements.txt
asgiref==3.4.1
Django==3.2.6
Pillow==8.3.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
